reafactored file structure...again, but better
