--- client/__init__.py
@@ -1,2 +1,0 @@
-from client.cli import CLI
-
