--- core/blocks/__init__.py
@@ -1,2 +1,0 @@
-import core.blocks
-import core.blocks.merkle
