--- core/transaction/__init__.py
@@ -1,1 +1,0 @@
-from core.transaction.transaction import Transaction
