+++ main.py
@@ -1,5 +1,5 @@
 ''' Main File '''
 # from block import Block
-from client import CLI
+from src.client import CLI
 
 CLI().cmdloop()
