+++ src/client/__init__.py
@@ -1,0 +1,2 @@
+from src.client.cli import CLI
+
