+++ src/client/helpers.py
@@ -1,0 +1,29 @@
+"""
+A file containing random functions that may or may not be useful.
+"""
+
+def cromulon():
+    """
+    Returns a cromulon
+    :return: a cromulon ascii img
+    """
+    ascii_img = '''
+                    ___          
+                . -^   `--,      
+               /# =========`-_   
+              /# (--====___====\ 
+             /#   .- --.  . --.| 
+            /##   |  * ) (   * ),
+            |##   \    /\ \   / |
+            |###   ---   \ ---  |     _______________________________
+            |####      ___)    #|   /                                |
+            |######           ##|  /  I like what you got. Good Job. |
+             \##### ---------- /  /                                  |
+              \####           (  <___________________________________|
+               `\###          |  
+                 \###         |  
+                  \##        |   
+                   \###.    .)   
+                    `======/  
+                    '''
+    return ascii_img
