+++ src/configuration/__init__.py
@@ -1,0 +1,1 @@
+from src.configuration.configuration import Configuration
