+++ src/configuration/singleton.py
