+++ src/core/__init__.py
@@ -1,0 +1,1 @@
+from src.core.api import mine, create_transaction, get_block, init_configuration
