+++ src/wallet/__init__.py
@@ -1,0 +1,1 @@
+from src.wallet.wallet import get_private_key, get_public_key, import_public_key
